backend_project/migrator/requirements.txt
mysql-connector-python==8.4.0
sqlparse==0.5.3
